Urgency Analysis

Urgency Analysis

get_urgency(v, subjects)

Arguments

v

Text vector or annotated data frame.

subjects

A vector or a list of subjects and related terms. If not declared, package will try to automatically detect subjects and related terms to these subjects. If users declare a vector, each element is treated as a independent topic and related terms to each of the elements will be automatically identified if possible. If users declare a list of subjects and related terms, function understands these as the topics and the related terms.

Value

A scored data frame.

Examples

# \donttest{
get_urgency(US_News_Conferences_1960_1980[1:10,3])
#> ✔ Extracted promises.
#> ✔ Extracted subjects.
#> ✔ Extracted similar topics for subjects.
#> ℹ Coding urgency components...
#> # A tibble: 175 × 20
#>    doc_id sentence_id      ntoken sentence poss  tags  lemmas adverbs adjectives
#>    <chr>  <chr>             <int> <chr>    <chr> <chr> <chr>    <dbl>      <dbl>
#>  1 text7  29                   35 We must… PRON… PRP … we mu…       0        2  
#>  2 text4  56 - 59              67 ,But — … PUNC… , : … ,but …       0        4  
#>  3 text5  177                  37 ,So , i… PUNC… , , … ,so ,…       0        2.4
#>  4 text3  84                   43 I think… PRON… PRP … i thi…       0        3.2
#>  5 text9  155 - 157 - 158…     84 And the… CCON… CC E… and t…       0        3.6
#>  6 text7  37 - 45 - 46         79 It is u… PRON… PRP … it be…       0        4.2
#>  7 text1  44                   25 It has … PRON… PRP … it ha…       0        0.6
#>  8 text8  69                   30 And I t… CCON… CC P… and i…       0        1  
#>  9 text1  130                   9 I 'll j… PRON… PRP … i wil…       0        1.6
#> 10 text9  116 - 117 - 120      68 ,I have… PROP… NNP … ,i ha…       0        1.2
#> # ℹ 165 more rows
#> # ℹ 11 more variables: nouns <chr>, entities <chr>, seg_id <chr>,
#> #   problem <chr>, promises <chr>, topic <list>, frequency <dbl>, timing <dbl>,
#> #   degree <dbl>, commit <dbl>, urgency <dbl>
get_urgency(US_News_Conferences_1960_1980[1:10,3],
            subjects = c("war", "inflation"))
#> ✔ Extracted promises.
#> ℹ Failed to identify related terms, subjects will be used to code topics.
#> ℹ Coding urgency components...
#> Error in get_urgency(US_News_Conferences_1960_1980[1:10, 3], subjects = c("war",     "inflation")): object 'similar_words' not found
get_urgency(US_News_Conferences_1960_1980[1:10,3],
            subjects = list("war" = c("war", "military", "guns"),
                            "inflation" = c("inflation", "interest rates", "prices")))
#> ✔ Extracted promises.
#> ℹ Coding urgency components...
#> # A tibble: 175 × 20
#>    doc_id sentence_id      ntoken sentence poss  tags  lemmas adverbs adjectives
#>    <chr>  <chr>             <int> <chr>    <chr> <chr> <chr>    <dbl>      <dbl>
#>  1 text7  29                   35 We must… PRON… PRP … we mu…       0        2  
#>  2 text4  56 - 59              67 ,But — … PUNC… , : … ,but …       0        4  
#>  3 text5  177                  37 ,So , i… PUNC… , , … ,so ,…       0        2.4
#>  4 text3  84                   43 I think… PRON… PRP … i thi…       0        3.2
#>  5 text9  155 - 157 - 158…     84 And the… CCON… CC E… and t…       0        3.6
#>  6 text7  37 - 45 - 46         79 It is u… PRON… PRP … it be…       0        4.2
#>  7 text1  44                   25 It has … PRON… PRP … it ha…       0        0.6
#>  8 text8  69                   30 And I t… CCON… CC P… and i…       0        1  
#>  9 text1  130                   9 I 'll j… PRON… PRP … i wil…       0        1.6
#> 10 text9  116 - 117 - 120      68 ,I have… PROP… NNP … ,i ha…       0        1.2
#> # ℹ 165 more rows
#> # ℹ 11 more variables: nouns <chr>, entities <chr>, seg_id <chr>,
#> #   problem <chr>, promises <chr>, topic <list>, frequency <dbl>, timing <dbl>,
#> #   degree <dbl>, commit <dbl>, urgency <dbl>
# }