Simulating urgency in priorities

Simulating urgency in priorities

Usage

sim_urgency(urgency, commitment, intensity, timing, frequency, pronoun = "We")

Arguments

urgency

Desired urgency score, optional.

commitment

Desired commitment score, optional.

intensity

Desired intensity score, optional.

timing

Desired timing score, optional.

frequency

Desired frequency score, optional.

pronoun

How would you like the simulated priorities to start? By default, priorities start with the pronoun "We".

Value

A sentence that matches the urgency or urgency dimension scores.

Details

Users can declare a score for one or more of the urgency dimensions or an urgency score. This means, if users may not declare an urgency score and the score for one or more dimensions at once. In those cases, the urgency score is favored.

Examples

# \donttest{
sim_urgency()
#> Urgency score: 1 
#> [1] "We do this."
sim_urgency(urgency = 0.5)
#> Urgency score: 0.5000087
#> [1] "We urgently would do this gradually."
sim_urgency(urgency = 2.5)
#> Urgency score: 2.50015
#> [1] "We vigorously must do this shortly."
sim_urgency(urgency = -2.5)
#> Urgency score: 2.50015
#> [1] "We vigorously must never do this shortly."
sim_urgency(commitment = 0.6)
#> Urgency score: 0.6 
#> [1] "We intend to do this."
sim_urgency(commitment = 0.6, intensity = 1.4)
#> Urgency score: 0.84 
#> [1] "We especially promise to do this."
sim_urgency(commitment = 0.6, intensity = 1.4, timing = 1.4)
#> Urgency score: 1.176 
#> [1] "We especially commit to do this early."
sim_urgency(commitment = 0.6, intensity = 1.2, timing = 1.4, frequency = 1.8)
#> Urgency score: 1.8144 
#> [1] "We substantially intend to do this early constantly."
# }